% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickCoefPlot.R
\name{QuickCoefPlot}
\alias{QuickCoefPlot}
\alias{qcp}
\title{QuickCoefPlot}
\usage{
QuickCoefPlot(model, iv.vars.names, plot.title, include.only, robust.se,
  cluster, cluster.vars.names, boot.se, boot.b, boot.plot.est, plot.lines,
  plot.margin, legend.on, colors.off, text.size, hide.summary.lines, xlim,
  horserace)
}
\arguments{
\item{model}{Fitted "lm" object.}

\item{iv.vars.names}{(Optional) Vector of desired independent variable names in table output (e.g. c("GDP per capita", "Population")). Defaults to values in "iv.vars" if none provided.}

\item{plot.title}{(Optional) Specifies the title of the coefficient plot. Defaults to no title.}

\item{include.only}{(Optional) Vector of coefficients by number to keep in the plot (e.g. c(1, 2, 6)). If none specified, defaults to all.}

\item{robust.se}{(Optional) If TRUE, returns robust standard errors calculated using a sandwich estimator from the "sandwich" package. Defaults to TRUE (i.e. robust rather than normal standard errors).}

\item{cluster}{(Optional) Name of variable by which cluster-robust standard errors should be computed using the cluster.vcov command of the multiwayvcov package. If this variable is not in the model, a data frame common to both the model variables and the clustering variable must be supplied.}

\item{cluster.vars.names}{(Optional) Desired name or label of clustering variable to be reported in table output (e.g. "Country" yields a note on the bottom of the table reading "Country-Clustered Standard Errors in Parenthesis"). If cluster specified but no "cluster.vars.names" provided, "Cluster-Robust Standard Errors in Parenthesis" is reported.}

\item{boot.se}{(Optional) If TRUE, calculates confidence interval based on bootstrap simulations. By default based on 100 such simulations, see boot.b below.}

\item{boot.b}{(Optional) Integer specifying the number of bootstrap simulation to run if boot.se is set to TRUE. Defaults to 100.}

\item{boot.plot.est}{(Optional) If TRUE, plots the bootstrap estimates semi-transparently on the coefficient plot. Defaults to FALSE.}

\item{plot.lines}{(Optional) If FALSE, does not plot confidence intervals on the coefficient plot. Defaults to TRUE.}

\item{plot.margin}{(Optional) Vector of plot margins in centimeters. Defaults to (1, 1, 1, 1).}

\item{legend.on}{(Optional) If TRUE turns on the plot legend. Defaults to FALSE.}

\item{colors.off}{(Optional) If TRUE turns off color scheme (sky-blue if p > 0.1, dark blue if p < 0.1, black if p < 0.05). Defaults to FALSE.}

\item{text.size}{(Optional) Text size for plot elements.}

\item{hide.summary.lines}{(Optional) Vector of summary lines to hide in plot output. If none supplied, defaults to none.}

\item{xlim}{(Optional) Vector of limits on x-axis of plot. If none supplied, this is automatically selected by ggplot.}

\item{horserace}{(Optional) If TRUE Produces a table comparing t-statistics instead of a coefficient plot.}
}
\description{
QuickCoefPlot (aka qcp) is an easy interface for linear regression coefficient plots in R. This includes the option to request robust and clustered standard errors, automatic labeling, and easy selection of coefficients to plot.
Written by Sondre U. Solstad (ssolstad@princeton.edu) - Please cite my github: github.com/sondreus/QuickCoefPlot.
}
\examples{

QuickCoefPlot(lm(x ~y, data = mydata))
qcp(lm(x ~ y, data = mydata, boot.se = TRUE, boot.plot.est = TRUE, legend.on = TRUE))

Please see: github.com/sondreus/QuickCoefPlot
}
\keyword{LS}
\keyword{cluster}
\keyword{coefplot}
\keyword{horse-race}
\keyword{lm}
\keyword{reg}
\keyword{regression}
\keyword{robust}
\keyword{robust.se}
\keyword{tstat}

